Task70:Bid validation
